+++ tests/integration/tx/context_test.go
@@ -3,29 +3,19 @@ package tx
 import (
 	"testing"
 
-	"github.com/stretchr/testify/require"
-	"google.golang.org/protobuf/proto"
-
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/log"
 	"cosmossdk.io/x/tx/signing"
-
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
+	"github.com/cosmos/cosmos-sdk/tests/integration/tx/internal"
 	"github.com/cosmos/cosmos-sdk/tests/integration/tx/internal/pulsar/testpb"
 	"github.com/cosmos/cosmos-sdk/testutil/configurator"
 	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
+	"github.com/stretchr/testify/require"
 )
 
-func ProvideCustomGetSigners() signing.CustomGetSigner {
-	return signing.CustomGetSigner{
-		MsgType: proto.MessageName(&testpb.TestRepeatedFields{}),
-		Fn: func(msg proto.Message) ([][]byte, error) {
-			testMsg := msg.(*testpb.TestRepeatedFields)
-			// arbitrary logic
-			signer := testMsg.NullableDontOmitempty[1].Value
-			return [][]byte{[]byte(signer)}, nil
-		},
-	}
+func ProvideCustomGetSigner() signing.CustomGetSigner {
+	return internal.TestRepeatedFieldsSigner
 }
 
 func TestDefineCustomGetSigners(t *testing.T) {
@@ -40,7 +30,7 @@ func TestDefineCustomGetSigners(t *testing.T) {
 				configurator.ConsensusModule(),
 			),
 			depinject.Supply(log.NewNopLogger()),
-			depinject.Provide(ProvideCustomGetSigners),
+			depinject.Provide(ProvideCustomGetSigner),
 		),
 		&interfaceRegistry,
 	)